feat: move all nav groups to a single component and remove redundant files
