--- src/components/nav-section.tsx
@@ -1,40 +1,0 @@
-"use client";
-
-import type { SidebarNavDataType } from "@/lib/types";
-
-import { Settings2 } from "lucide-react";
-
-import {
-  SidebarGroupLabel,
-  SidebarMenuButton,
-  SidebarMenuItem,
-  SidebarGroup,
-  SidebarMenu,
-} from "@/components/ui/sidebar";
-
-import Link from "next/link";
-
-
-const menu: SidebarNavDataType = [
-  { title: "Settings", url: "/dashboard/settings", icon: Settings2 },
-];
-
-export function NavSubMenu() {
-  return (
-    <SidebarGroup className="group-data-[collapsible=icon]:hidden">
-      <SidebarGroupLabel>Others</SidebarGroupLabel>
-      <SidebarMenu>
-        {menu.map((item) => (
-          <SidebarMenuItem key={item.title}>
-            <SidebarMenuButton asChild>
-              <Link href={item.url}>
-                <item.icon />
-                <span>{item.title}</span>
-              </Link>
-            </SidebarMenuButton>
-          </SidebarMenuItem>
-        ))}
-      </SidebarMenu>
-    </SidebarGroup>
-  );
-}
